latest

+++ admin/php/deleteUser.php
@@ -1,8 +1,9 @@
 <?php
+    session_start();
     include('connection.php');
     $con = connect();
 
-    if(isset($_POST)){
+    if(isset($_POST) && isset($_SESSION['admin_id'])){
         $userId = $_POST['user_id'];
         $query = "DELETE FROM users WHERE id = '$userId'";
         $con->query($query) or die($con->error);
@@ -12,5 +13,7 @@
 
 
         echo 'ok';
+    }else{
+        echo 'Unauthorized';
     }
 ?>
